Fix id tracking.
When assembling the highlight markers, the global `RangeHighlighter.id`
is set on the query for each search string inside the query. Instead,
it should be set once at the beginning.

Also, the `do` function (which actually applies the CSS classes to the
highlight) needs to know the search string index in the context of the
query to produce the correct CSS class name.