src: fix whitespace bug
The problem seems to have been caused by the fact that the presentation
of the string does not match its internal representation: the string `a
b` is shown as `a b` with one space between `a` and `b`.  This has
been **fixed** by accounting for specific cases where a selection's
start and end elements are the same and using the selection's start and
end offsets instead of its text length.