Rework eslint invocation to avoid needless warnings
Specifying exact directories in which one or more files are found to be
excluded by the special `.eslintignore` directives file causes eslint to
produce a warning per exclusion.  However, if one specifies `.` as the path to
process - as opposed to specific directories/files - eslint does not produce
any warnings about excluded files.