test: initial support for full document tests
Full document tests are tests where the HTML view being used as the basis for
the tests represents a full document rather than a snippet that is appended to
a container.  In addition, the HTML Highlighter is instantiated with its
container set to `document`, rather than the default `document.body`.