Add master index file
When running a webpack dev server build via `yarn run start` and navigating to
the root URL of the server (e.g. `http://localhost:8080`), it can be confusing
for users to know what to do next, as a directly listing of the project's
repository is rendered.  The master index file contained in this commit solves
this by automatically redirecting the user to the main (and currently only)
example in `examples/monolith/`, thus providing an immediate starting point.